requirements.txt
openai==1.58.1
arxiv==2.1.0
requests==2.31.0
python-dotenv==1.0.0
